requirements.txt
gensim
bs4
Mastodon.py
nltk
numpy
pandas
re
langdetect
